Extract extra text to a HOC
