+++ src/SignupWithMobile.js
@@ -12,6 +12,7 @@ import TextInput from "./TextInput";
 import type { Props as TextInputProps } from "./TextInput";
 import CountryPicker from "./CountryPicker";
 import type { Props as CountryPickerProps } from "./CountryPicker";
+import type { Props as ExtraTextProps } from "./withExtraText";
 
 const defaultStyles = StyleSheet.create({
   header: {
@@ -115,8 +116,8 @@ export type Props = {
   title?: string,
   titleStyle?: TextStyle,
 
-  countryPickerProps?: CountryPickerProps,
-  mobileNumberProps?: TextInputProps,
+  countryPickerProps?: ExtraTextProps<CountryPickerProps>,
+  mobileNumberProps?: ExtraTextProps<TextInputProps>,
 
   submitButtonStyle?: ViewStyle,
   submitButtonText?: string,
