+++ src/TextInput.js
@@ -1,8 +1,9 @@
 // @flow
 import * as React from "react";
-import { View, TextInput, StyleSheet } from "react-native";
+import { View, TextInput as NativeTextInput, StyleSheet } from "react-native";
 
 import Text from "./Text";
+import withExtraText from "./withExtraText";
 import { TextStyle, ViewStyle } from "./styles";
 
 const defaultStyles = StyleSheet.create({
@@ -11,66 +12,24 @@ const defaultStyles = StyleSheet.create({
     padding: 0,
     margin: 0,
   },
-  extraText: {
-    alignSelf: "flex-start",
-  },
-  nonvisible: {
-    opacity: 0,
-  },
 });
-
-function makeExtraText(
-  error?: string | null,
-  errorStyle?: TextStyle,
-  option?: string,
-  optionStyle?: TextStyle
-): Text | null {
-  if (error) {
-    return <Text style={[defaultStyles.extraText, errorStyle]}>{error}</Text>;
-  } else if (option) {
-    return <Text style={[defaultStyles.extraText, optionStyle]}>{option}</Text>;
-  } else if (error !== undefined) {
-    // make sure layout won't unstable, whether error message show or not.
-    return (
-      <Text
-        style={[defaultStyles.extraText, errorStyle, defaultStyles.nonvisible]}
-      >
-        nonvisible
-      </Text>
-    );
-  }
-}
 
 export type Props = {
   style?: TextStyle,
-  error?: string | null,
-  errorStyle?: TextStyle,
-  option?: string,
-  optionStyle?: TextStyle,
-  containerStyle?: ViewStyle,
 };
 
-// $FlowFixMe - `React.forwardRef` is not defined in Flow, yet.
-export default React.forwardRef((props: Props, ref?) => {
-  const {
-    error,
-    option,
-    errorStyle,
-    optionStyle,
-    style,
-    containerStyle,
-    ...rest
-  } = props;
+function TextInput(props: Props, ref?) {
+  const { style, ...rest } = props;
 
   return (
-    <View style={containerStyle}>
-      <TextInput
-        {...rest}
-        ref={ref}
-        style={[defaultStyles.textInput, style]}
-        underlineColorAndroid="transparent"
-      />
-      {makeExtraText(error, errorStyle, option, optionStyle)}
-    </View>
+    <NativeTextInput
+      {...rest}
+      ref={ref}
+      style={[defaultStyles.textInput, style]}
+      underlineColorAndroid="transparent"
+    />
   );
-});
+}
+
+// $FlowFixMe
+export default withExtraText<Props>(React.forwardRef(TextInput));
